+++ wazo_phoned/plugins/yealink/services.py
@@ -31,6 +31,29 @@ class YealinkService:
                 else:
                     self._send_notify(endpoint, 'DNDOff')
 
+    def update_forward_unconditional(self, user_uuid, destination, enabled):
+        try:
+            self.confd.users(user_uuid).update_forward('unconditional', {
+                'destination': destination,
+                'enabled': enabled,
+            })
+        except RequestException as e:
+            response = getattr(e, 'response', None)
+            status_code = getattr(response, 'status_code', None)
+            if status_code == 404:
+                raise NoSuchUser(user_uuid)
+            raise
+
+    def notify_forward_unconditional(self, user_uuid, destination, enabled):
+        lines = self._find_lines(user_uuid)
+        for line in lines:
+            endpoint = line.get('name')
+            if endpoint:
+                if enabled:
+                    self._send_notify(endpoint, 'AlwaysFwdOn={}'.format(destination))
+                else:
+                    self._send_notify(endpoint, 'AlwaysFwdOff')
+
     def update_forward_busy(self, user_uuid, destination, enabled):
         try:
             self.confd.users(user_uuid).update_forward('busy', {